requirements.txt
bson==0.5.8
dataclasses==0.6
dj-static==0.0.6
Django==2.2.12
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gunicorn==20.0.4
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
static3==0.7.0
whitenoise==5.1.0